requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.6
aiofiles>=23.2.0
pgvector>=0.2.0
sentence-transformers>=2.2.0
transformers>=4.36.0
torch>=1.9.0
accelerate>=0.25.0
numpy>=1.24.0
pandas>=2.1.0
python-dotenv>=1.0.0
Pillow>=10.1.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-docx>=1.1.0
openpyxl>=3.1.0
pytesseract>=0.3.10
SpeechRecognition>=3.10.0
jinja2>=3.1.0